// lib/json-recipe.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Ingredient =
  | string
  | {
      /**
       * The quantity of the specified item, in terms of the specified unit.
       */
      quantity: Quantity;
      /**
       * The unit of measurement which the specified quantity uses (e.g. "cup").
       */
      unit: string;
      /**
       * The item of which this ingredient consists (e.g. "apple").
       */
      item: string;
      /**
       * How the ingredient should be prepared prior to use (e.g. "chopped").
       */
      preparation?: string | string[];
    };
export type Quantity = number | string;
/**
 * A single step in the preparation of the recipe.
 */
export type DirectionStep = string;

/**
 * A schema for recipes.
 */
export interface JsonRecipe {
  /**
   * Meta information that does not relate directly to the recipe itself, such as the schema version used.
   */
  meta?: {
    /**
     * The version of this schema under which the recipe was written. The schema follows semantic versioning, so recipes written under version 1.2.3 will be compatible with schema version 1.4.5, but not necessarily the other way around.
     */
    schemaVersion?: string;
    [k: string]: any;
  };
  /**
   * The title of the recipe.
   */
  title: string;
  /**
   * The source of the recipe (e.g. the person who wrote it and the book where it was found).
   */
  source?: {
    /**
     * The full name of the author of the recipe.
     */
    author: string;
    /**
     * The location from which the recipe was retrived (e.g. a website or a book).
     */
    location?: {
      /**
       * The URL of the source material.
       */
      url: string;
      /**
       * The date on which the recipe was retrieved from the source.
       */
      retrievalDate?: string;
      [k: string]: any;
    };
    [k: string]: any;
  };
  /**
   * The ingredients of the recipe.
   */
  ingredients: (
    | Ingredient
    | {
        /**
         * The heading for this group of ingredients.
         */
        heading: string;
        /**
         * The ingredients under this heading.
         */
        ingredients: Ingredient[];
      })[];
  /**
   * The instructions for preparing the finished product from the ingredients.
   */
  directions: (
    | DirectionStep
    | {
        /**
         * The heading for this group of directions.
         */
        heading: string;
        /**
         * The directions under this heading.
         */
        directions: DirectionStep[];
      })[];
  [k: string]: any;
}